requirements.txt
pillow>=8.3.2
requests>=2.26.0
python-dotenv>=0.19.0
django>=3.2.0
filetype
standard-imghdr
rembg
onnxruntime
torch
torchvision
u2net
opencv-python>=4.5.5
numpy>=1.21.0
gdown>=4.7.1
fitz
python-docx
PyMuPDF
openpyxl
pandas